✨ Feat: UserPoint 사용 기능 구현 및 단위테스트

+++ src/main/kotlin/kr/hhplus/be/server/user/application/UserPointResult.kt
@@ -1,16 +1,17 @@
 package kr.hhplus.be.server.user.application
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 class UserPointResult {
     data class Charge(
         val userId: Long,
-        val pointAfterCharge: Long,
+        val pointAfterCharge: BigDecimal,
         val updatedAt: LocalDateTime?
     )
     data class Retrieve(
         val userId: Long,
-        val point: Long,
+        val point: BigDecimal,
         val updatedAt: LocalDateTime?
     )
 }
